Extract the individual screens used for screenshots
